% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_highlight}
\alias{df_highlight}
\title{Mappings of PheCodes to ICD codes.}
\format{
A data frame with 106393 rows and 6 variables:
\describe{
  \item{ICD_version}{ICD version, ICD-9 or ICD-10-cm}
  \item{Phecode}{ICD-9 and ICD-10-cm Codes}
}
}
\source{
\url{https://phewascatalog.org/}
}
\usage{
df_highlight
}
\description{
A dataset containing the mappings of PheCodes to ICD-9 and ICD-10-cm Codes.
}
\keyword{datasets}
