% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_icd}
\alias{dict_icd}
\title{Information of ICD codes}
\format{
A data frame with 99765 rows and 3 columns:
\describe{
  \item{id}{ICD code}
  \item{term}{ICD string}
  \item{version}{ICD version, ICD-9 or ICD-10-cm}
}
}
\source{
\url{https://uts.nlm.nih.gov/uts/umls/}
}
\usage{
dict_icd
}
\description{
A datatable of the information of icd codes from UMLS version 2021AB
}
\keyword{datasets}
